Re-implemented via filter and added largest_first setting
Re implemented sweep_until as a filter. As that is what it is, so there are now minimal changed to the engine.

Updated Filter to be extendable, so filters can get a reference to the engine instance.

Added largest_first so that if users prefer (i do), If an item is eligible for deletion, delete the biggest item first - as it leads to less media being removed.
